+++ Colors.swift
@@ -1,0 +1,15 @@
+//
+//  Colors.swift
+//  Virtusize
+//
+//  Created by Kuei-Jung Hu on 2021/01/21.
+//  Copyright © 2021 Virtusize AB. All rights reserved.
+//
+
+import Foundation
+
+/// The Virtusize themed colors used in the Virtusize SDK
+extension UIColor {
+	static var vsGray900Color: UIColor { #colorLiteral(red: 0.09803921569, green: 0.09803921569, blue: 0.09803921569, alpha: 1) }
+	static var vsTealColor: UIColor { #colorLiteral(red: 0.2431372549, green: 0.8235294118, blue: 0.7294117647, alpha: 1) }
+}