+++ Source/UI/VirtusizeViewStyle.swift
@@ -1,5 +1,5 @@
 //
-//  Localization.swift
+//  VirtusizeViewStyle.swift
 //
 //  Copyright (c) 2018-present Virtusize KK
 //
@@ -22,26 +22,8 @@
 //  THE SOFTWARE.
 //
 
-import Foundation
-
-/// This class is used to localize texts in the SDK
-class Localization {
-
-    static let shared: Localization = Localization()
-
-	/// Localizes a text corresponding to a key
-	///
-	/// - Parameters:
-	///   - key: The key for a string in the table identified by tableName.
-	///   - language: Pass `VirtusizeLanguage` if you'd like to localize the text in a designated language
-	/// - Returns: A localized string based on the device's default language
-	func localize(_ key: String, language: VirtusizeLanguage? = nil) -> String {
-		let path = Bundle(for: Localization.self).path(
-			forResource: language?.rawValue ?? Virtusize.params?.language.rawValue,
-			ofType: "lproj") ?? Bundle(for: Localization.self).path(
-				forResource: VirtusizeLanguage.ENGLISH.rawValue,
-				ofType: "lproj"
-			)!
-		return Bundle(path: path)?.localizedString(forKey: key, value: nil, table: "VirtusizeLocalizable") ?? key
-	}
+/// This enum contains all the styles that can be applied to the Virtusize view
+public enum VirtusizeViewStyle: String {
+	case BLACK = "Black"
+	case TEAL = "Teal"
 }